requirements.txt
pandas
matplotlib
python-dotenv
requests
fpdf2
